@@ -33,15 +33,11 @@ namespace iSukces.Code.Tests.EqualityGenerator
                 }
             }
 
-            public static bool operator ==(ClassWithEnumProperties1 left, ClassWithEnumProperties1 right)
-            {
-                return Equals(left, right);
-            }
+            public static bool operator ==(ClassWithEnumProperties1 left, ClassWithEnumProperties1 right) => 
+                Equals(left, right);
 
-            public static bool operator !=(ClassWithEnumProperties1 left, ClassWithEnumProperties1 right)
-            {
-                return !Equals(left, right);
-            }
+            public static bool operator !=(ClassWithEnumProperties1 left, ClassWithEnumProperties1 right) => 
+                !Equals(left, right);
 
         }
 
