@@ -42,15 +42,11 @@ namespace iSukces.Code.Tests.EqualityGenerator
                 }
             }
 
-            public static bool operator ==(ClassWithManyProperties left, ClassWithManyProperties right)
-            {
-                return Equals(left, right);
-            }
+            public static bool operator ==(ClassWithManyProperties left, ClassWithManyProperties right) => 
+                Equals(left, right);
 
-            public static bool operator !=(ClassWithManyProperties left, ClassWithManyProperties right)
-            {
-                return !Equals(left, right);
-            }
+            public static bool operator !=(ClassWithManyProperties left, ClassWithManyProperties right) => 
+                !Equals(left, right);
 
         }
 
