@@ -31,15 +31,9 @@ namespace iSukces.Code.Tests.EqualityGenerator
                 }
             }
 
-            public static bool operator ==(ClassWithNullable left, ClassWithNullable right)
-            {
-                return Equals(left, right);
-            }
+            public static bool operator ==(ClassWithNullable left, ClassWithNullable right) => Equals(left, right);
 
-            public static bool operator !=(ClassWithNullable left, ClassWithNullable right)
-            {
-                return !Equals(left, right);
-            }
+            public static bool operator !=(ClassWithNullable left, ClassWithNullable right) => !Equals(left, right);
 
         }
 
